requirements.txt
dj-database-url==0.5.0
Django==2.2.4
django-ckeditor==5.7.1
django-js-asset==1.2.2
django-taggit==1.1.0
gunicorn==20.0.4
Markdown==3.1.1
Pillow==6.1.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
